refactoring: move get_region function
